Now pruning/creating obstacles on collision (trigger really).

+++ Assets/Scripts/TriggerColliders.cs
@@ -3,24 +3,37 @@ using System.Collections;
 
 public class TriggerColliders : MonoBehaviour {
 	public ObstacleManager m_obstacleManager;
+	public bool m_create = false;
+	public bool m_destroy = false;
 
-	void OnTriggerEnter2D(Collider2D other) {
-		switch (this.tag) {
-			case "createCollide":	
-				Debug.Log ("Creating new obstacle on createCollide");
-				m_obstacleManager.createObstacle ();
-				break;
-		}
-		Debug.Log ("collision start");
+	void start() {
 	}
 
+	/**
+	 * A bit hacky, but basically this is triggering twice,
+	 * beacuse our 'obstacle' is composed of 2 different colliders (top & bottom).
+	 */
 	void OnTriggerExit2D(Collider2D other) {
 		switch (this.tag) {
+			case "createCollide":
+				if (m_create) {
+					Debug.Log ("Creating new obstacle on createCollide");
+					m_obstacleManager.createObstacle ();
+					m_create = false;
+				} else {
+					m_create = true;
+				}
+				break;	
+
 			case "destroyCollide":
-				Debug.Log ("Creating new obstacle on createCollide");
-				m_obstacleManager.destroyObstacle (other);
+				if (m_destroy) {
+					Debug.Log ("Destroying new obstacle on createCollide");
+					m_obstacleManager.destroyObstacle (other);
+					m_destroy = false;
+				} else {
+					m_destroy = true;
+				}
 				break;
 		}
-		Debug.Log ("collision exit");
 	}
 }
